[Math] Cleanup

+++ Sources/NexusFramework/Math/Rotation.h
@@ -7,7 +7,7 @@
 #include "NexusFramework/Types/Strings/String.h"
 #include "NexusFramework/Types/Strings/StringFunctions.h"
 
-#include "NexusFramework/Math/Types/Vector.h"
+#include "NexusFramework/Math/Vector.h"
 
 namespace NxFr
 {
@@ -98,8 +98,8 @@ namespace NxFr
 
 		bool operator==(Euler Other) const { return x == Other.x && y == Other.y && z == Other.z; }
 		bool operator!=(Euler Other) const { return !(*this == Other); }
-		float& operator[](uint64 Index) { return Index == 0 ? x : Index == 1 ? y : Index == 2 ? z : x; }
-		const float& operator[](uint64 Index) const { return Index == 0 ? x : Index == 1 ? y : Index == 2 ? z : x; }
+		float& operator[](uint8 Index) { return Index == 0 ? x : Index == 1 ? y : Index == 2 ? z : x; }
+		const float& operator[](uint8 Index) const { return Index == 0 ? x : Index == 1 ? y : Index == 2 ? z : x; }
 
 	public:
 		float x, y, z;
@@ -206,8 +206,8 @@ namespace NxFr
 		Quaternion& operator=(Quaternion Other) { x = Other.x; y = Other.y; z = Other.z; w = Other.w; return *this; }
 		bool operator==(Quaternion Other) const { return x == Other.x && y == Other.y && z == Other.z && w == Other.w; }
 		bool operator!=(Quaternion Other) const { return !(*this == Other); }
-		float& operator[](uint64 Index) { return Index == 0 ? x : Index == 1 ? y : Index == 2 ? z : Index == 3 ? w : x; }
-		const float& operator[](uint64 Index) const { return Index == 0 ? x : Index == 1 ? y : Index == 2 ? z : Index == 3 ? w : x; }
+		float& operator[](uint8 Index) { return Index == 0 ? x : Index == 1 ? y : Index == 2 ? z : Index == 3 ? w : x; }
+		const float& operator[](uint8 Index) const { return Index == 0 ? x : Index == 1 ? y : Index == 2 ? z : Index == 3 ? w : x; }
 
 		Quaternion& operator*=(Quaternion Other) { *this = RotationUtility::Concatenate(*this, Other); return *this; }
